# Use `names<-` on base types in `vec_set_names()`